requirements.txt
Django==5.2.1
django-allauth==0.61.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
djangorestframework==3.15.1
django-filter==24.2
django-debug-toolbar==4.3.0
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0 